Added attributes and methods for Bus struct
These include reset, inc for each clock cycle, read, and just the
declerations for write and load program. Write needs to implement mirroring
and for load_program I need to understand the cartridge memory map a bit better
to implement. Also need to think about how to implement the file reading and how
that will be triggered in the main loop.